modifief graph.js and Graph.test.js
