+++ js/components/graph/__tests__/Graph.test.js
@@ -1,6 +1,9 @@
+import React from "react"
 import TestGraph from "./TestGraph"
 import { fireEvent } from "@testing-library/react"
 import { ZOOM_INCREMENT, KEYBOARD_PANNING_INCREMENT } from "../Graph"
+import Graph from "../Graph"
+import { shallow } from "enzyme"
 
 describe("Graph Navigation", () => {
   it("Should pan right when the right arrow key is pressed", async () => {
@@ -138,5 +141,44 @@ describe("Graph Navigation", () => {
     let expectedDims = initialDims.map(dim => dim * (1 + ZOOM_INCREMENT))
     expect(newDims[0]).toBe(expectedDims[0])
     expect(newDims[1]).toBe(expectedDims[1])
+  })
+})
+
+describe("Graph rendering RegionGroup", () => {
+  it("should match shallow snapshot", async() => {
+    
+    const props = {
+      regionsJSON: [
+        {
+          graph: 1,
+          points: [
+            [17.386348, 281.07376883],
+            [17.386348, 14.41568908],
+            [316.385978, 14.41576883],
+            [316.386348, 281.07376883],
+          ],
+          isRegion: true,
+          stroke: "",
+          fill: "#6276b9",
+          id_: "p83",
+          source: "",
+          target: "",
+        },
+      ],
+      labelsJSON: [
+        {
+          graph: 1,
+          rId: "tspan4346-9",
+          text: "Systems",
+          pos: [1088.1677413999998, 201.36453113000002],
+          fill: "#000000",
+          align: "begin",
+        },
+      ],
+    }
+    // const graph = await TestGraph.build()
+    // let wrapper = graph.getByTestId("regions")
+   
+    // expect(wrapper).toMatchSnapshot()
   })
 })
